Add SSTNetwork class; add Qt dependencies to build.
The real implementation of the SSTNetwork class is in its own file since it needs to be a QObject and having it separated makes the build setup (moc stuff) easier. The "net" option controls which network layer to use, currently still defaulting to raknet.  Add a service call, which defaults to doing nothing, to networking; this allows a separate event loop to be handled internally (Qt in the case of SST).  The SST implementation currently uses a separate connection for each direction. We may want to change this in the future, but it simplifies the initial connection setup by not having potentially simultaneous, identical connection requests; a simple scheme to select one, or to setup all connections before starting a run, would also work.