Shift the majority of the work for proximity into a separate thread.  It should be low cost, i.e. it only updates periodically (for relatively long periods, currently 100ms) and sleeps the rest of the time.  Most of the work now done in the main thread for Proximity::service is simply shuffling data back and forth from the main thread.
There are currently 2 known problems with this code.  First, subscriptions in AlwaysUpdateLocationPolicy seem to get inconsistent and this would happen, presumably, with any implementation.  The source of the problem is unkown, but it currently doesn't seem to affect results.  The affected lines are marked with FIXMEs.

Second, in order to avoid having locks on all the queries and the map which maintains them, we need to maintain (obj,query angle) information in the main thread too.  This is undesirable since we maintain the same information in two places, but works for now.