Created a JSPositionListener subclass for JSPresenceStruct and JSVisibleStruct.  This class should take care of automatically listening for updates from the proxy managers of the hosted object.  Having an underlying positionlistener subclass allowed me to delete a lot of extraneous code from jsobjectscript (eg. getPosition, getVelocity, etc.).
Also, created a JSVisibleStructMonitor from which JSObjectScript inherits.  The rough idea is that we don't want to make many copies of JSVisibleStructs that point to the same proxy object and which are observed by the same internal presence.  VisibleStructMonitor manages all visiblestructs and pushes updates to those that aren't necessarilly being watched from a local presence from ones that are.
With JSVisibleStructMonitor, removed the visible array system.visible.  Scripters are responsible for keeping track of the visible objects that they've seen themselves.

Finally, made a couple of changes to testVisible.em to actually make sure that all of these changes worked.