Fix crash when own presence is removed and re-added to pinto results.
When the presence used for an Ogre camera was removed and re-added to
pinto results, it caused the system to get into a bad state where it
didn't track the ProxyObject anymore but still had the camera. Fix
this by allowing the camera to reparent to the new ProxyObject. This
also requires carefully tracking state to make sure we can always
return values for goal position and orientation, as well as that we
reset to the same camera mode and settings.