Add sirikata.ui for safer wrapping of UI modules.
Add sirikata.ui function which wraps a UI module's code, contained in
a function, and associates it with a name. It automatically wraps
sirikata.event to prefix the UI module's name to try to better avoid
conflicts. The name should match up with the name provided to the
Emerson std.graphics.GUI class (and the addGUIModule and related
methods). Fixes #222.