Add physics properties UI.
Add a physics properties UI and load it for the defaultAvatar. In
doing so, add support for setting physics on Movable objects, fix some
styling issues that leaked beyond the components they were used for,
and update sirikata.ui.window to be a constructor and provide a more
useful interface for working with the window.