Make all storage operations async.
Make all storage operations asynchronous, either taking a callback
directly if they are the only item in the transaction, or using a
unified CommitCallback signature which now includes read
results. Adjust the Emerson API for restoring objects to use the new
interface, which in turn makes the Emerson API asynchronous, using a
callback.  Fix many test scripts that used the old form, and a few
that had also fallen out of date.