Reorganize tests into their own directory.
Since the tests binary is really standalone and separate from any
given library it is testing, reorganize tests into a separate
directory. It is treated as a standalone binary, links in the
necessary libraries, and is really a normal client of the other
libraries rather than having their tests embedded in each library.