Make ProxyEntity handle HostedObject destroying and recreating ProxyObjects in response to Proximity events.
Make ProxyEntity (and OgreSystem) capable of switching between
ProxyObjects as they are destroyed and new versions created due to
Proximity events. Previously, even though we were trying to mask the
quick removal and readdition, the effect was disabled because the
ProxyObject triggered the creation of a new ProxyEntity and the old
ProxyObject never got revalidated by the main loop. With this fix,
only one ProxyEntity is available for an object in the world and
therefore it gets all the necessary messages to learn of a deletion
and quick readdition. Fixes #283.