Fix loading of ProxyEntities for existing ProxyObjects.
We were sometimes seeing completely blank screens after loading up the
graphics system. It turns out that we've been getting Ogre fully
initialized *after* all the proximity results were received, meaning
it was working entirely off of existing ProxyObjects. This path was
not working correctly because we now rely not only on construction of
the ProxyEntity, but also that a validation call is received. Now we
trigger that explicitly and make sure that full, up-to-date
information is extracted from the ProxyObject when validation occurs.