Effectively revert 29b388c59 because it breaks the build in a number of ways.
First, the symlink isn't necessary if you aren't even building with
Ogre -- the check needs to be wrapped by checks for things that use
it. Second, JQuery *is not* required to run with graphics:
simplecamera doesn't use Berkelium at all and while the
defaultAvatar.em script uses Berkelium and jquery for UI, it is not
required.