Add basic raytrace support to libmesh and Emerson.
Add raytracing utilities for Meshdatas in libmesh. Expose this
functionality to Emerson as a single call for visibles and
presences. This raytrace is only on the mesh and doesn't take into
account transformations, so additional versions that take rays from
world space and which deal with raytracing from a different coordinate
system (i.e. a presence's) still need to be added to Emerson.