Recover from WebView crashes.
When a renderer process crashes, recover by reloading the page. For
the UI, we provide additional hooks to tell when the GUI was became
ready after reloading so that anything dynamically loaded GUIs can be
reloaded. Bound callbacks will still be valid, so usually the GUI only
needs to reinject the dynamic JS that gets loaded to reinitialize its
display. Fixes #286.