Add modelRotation setting on presences.
By wrapping some calls, we automatically add/remove it to the
orientation specified normally, so it passes the correct single value
to the system, but keeps the model orientation and 'regular'
orientation separate from the perspective of the script. If you use
this on the mesh tied to your camera you will end up with a different
orientation than expected, but this can be fixed when more of the
camera control is pushed into Emerson. Fixes #326.