Fix scaling for triangle meshes in Bullet.
This changes two things. First, separate the scaling down to a unit
sphere from the scale requested by the user. This should allow reuse
of meshes later on (since the data in the btTriangleMesh is now the
same across multiple btBvhTriangleMeshShapes). Second, fix the
application of scaling so it is applied as scale * instance_transform
instead of the reverse. The mesh-wide scaling needs to be applied
after instance transforms -- in the wrong order meshes with
non-trivial hierarchy and instancing behaved incorrectly.