Fix cleanup bugs which were causing crashes.
This change makes HostdObject and ObjectScript Services and triggers
their start() and stop() methods at the appropriate time. It also
rearranges some of the cleanup in Emerson scripts to properly handle
shutdown, removing a number of listeners which were left even after
deletion. Finally, it blocks actions in HostedObject after a stop is
requested so that HostedObject should not be responsible for any
further asynchronous operations.