Rework JSVisibleManager to properly track lifetimes based on ProxyObjects, visibles, and presences.
Rework the implementation of JSVisibleManager so it properly tracks
data over the lifetime of ProxyObjects, visibles, and presences. It
also now properly subscribes for updates from all ProxyObjects so that
it always has the most up to date information, instead of just the
information from the first copy of the ProxyObject to be found. Fixes