Making certain identifiers (like check resources function) un-overridable by user. Fixes #267.
Certain keywords shouldn't be overridable so that the user can't
get free cycles. The parser now matches and consumes keywords so that
they aren't treated as regular identifiers. This will result in either
a parser error or invalid JS. In the future, maybe it should emit
a better error message.