Request shutdown when last HostedObject dies.
Fix disconnection requests to clean presences out of the ObjectHost
and make the ObjectHost track whether this was the last presence,
requesting shutdown if no more HostedObjects exist to do more work. It
would be better if the IOService just ran out of work, but this is a
reasonable stopping condition for now.