Change ToCString to FromV8String and use it for string conversions.
Get rid of the relatively dangerous ToCString (which would never
account for string length, meaning it couldn't handle null characters
properly). Replace it with FromV8String which uses a real String
instead of char*, handles length properly, and only works for
converting UTF-8, which is what we should always be using. Get rid of
all uses of v8::String::AsciiValue so we know we're converting UTF-8
properly.