Add EZUI, courtesy Ben Christel.
Add EZUI, a library that makes it simple to create and interact with
UIs. UIs are fully specified in the object which want to present them,
but are executed in client objects by shipping them in messages and
then presenting them via the EZUI Viewer loaded by objects with
graphics support.