Fix warnings about computing normals for points, lines.
Don't warn about trying to compute normals for points, lines, and
linestrips since they don't need normals computed. Still generate a
warning if we get to the point of trying to compute them and we had
some to start, i.e. if the number of normals doesn't match the number
of points.

The existing setup for loading models would trigger this warning
because we'd try to append SubMeshGeometries and some would have
normals, some wouldn't. To fix this, separate triangulate from
reduce-draw-calls so that we can force triangulation, compute normals,
and then perform the remaining reduce-draw-calls operations. Those
operations should no longer fail due to mismatching normals for the
same primitive types, but could still fail if we try to merge, e.g.,
lines without normals with triangles with normals.