Make Proximity a general interface for query processors.
Move Proximity into libspace as a generic interface, with only
add and remove query, along with other utilities which must be
implemented for any implementation. Move AggeregateManager into
libspace so it is usable by implementations. Make the libprox
implementation a plugin in space-prox and use it as the default
implementation.