Fix request provided by Transfer callbacks to be the original.
Fix the TransferRequest provided by Transfer callbacks to be the
original request, even if the callback is triggered by another,
identical request. This was causing problems because some code would
rely on some information (e.g. the requested URI) to be the same as
the original request. Although it's more expensive this way, it's very
unintuitive to have the callback provide a different request.