Minimize VWObject interface.
Cut out everything which is not actually used via the VWObject
interface, which is everything except for the required interfaces and
a method to get the ProxyManager for a given presence. Generally,
VWObject and libproxyobject have been used only minimally and are
mostly observers (not controllers) of the state of the system, so we
shouldn't need this functionality expanded in the future.