Convert SessionManager from ODP::Service to OHDP::Service.
Convert SessionManager from ODP::Service to OHDP::Service since the
ODP::Service implementation was a hack to get the TimeSync code
incorporated without depending on the SessionManager
interface. OHDP::Service is what we really want from
SessionManager. Also convert the TimeSync code to use the interface
which both simplifies it (OHDP deals with converting to
ObjectMessages and getting the formats right) and gets what should
have been logic internal to TimeSyncClient in the right place and out
of SessionManager.