Initiate OHDP SST streams between OH and space as part of connection process.
As part of the connection process, initiate SST connections over OHDP
between the space and object host. The object host requires that this
connection must occur before the SpaceNodeConnection is considered
complete, so it is a required step in the connection process. NOTE:
This is a breaking protocol change: old object hosts are still able to
connect to new space servers (since the condition isn't enforced
there), but new object hosts will fail to connect to old space
servers.