Reorganize SpaceContext to use self-registration for components, add ObjectHostSessionManager to SpaceContext, and self-registered components to allow other components (LocationService, Proximity) to register themselves as listeners with session managers.
Make the base class of each component stored in Space perform
self-registration by making the interfaces friends and having their
constructors set the corresponding field in SpaceContext. Add
ObjectHostSessionManager with SpaceContext and rename the
ObjectSessionManager field to make it unambiguous. With these elements
registered earlier, make other components, namely LocationService and
Proximity, register themselves as listeners with session managers
instead of relying on Server to do it for them.