Make Factory throw an exception for missing constructors.
Since Factories are almost universally used by requesting a
constructor, using it's result directly, and not checking that a valid
value was returned, make Factory throw an exception when it fails to
find the requested constructor. This default behavior triggers an
error much earlier than what was previously happening -- segfaults due
to null pointers. If you want to just check the return value,
getConstructorOrDefault() can be used instead of getConstructor().