Add DIR_EXE_BUNDLE type to Path utilities.
The DIR_EXE_BUNDLE differs from DIR_EXE on OSX, where it finds the
parent path of the .app bundle, if the executable is in one in the
standard location. This is useful especially when searching for other
files based on the executable's location since it should make the
locations uniform across all the platforms.