Use Listener interface for orphan loc updates.
Use a Listener interface and simply request that
OrphanLocUpdateManager notify the listener of all pending orphan
updates. In contrast to the previous approach of just getting a list
of updates, this keeps the internals of OrphanLocUpdateManager
internal, avoids heap-allocating the LocUpdate since it doesn't have
to be returned to the caller, avoids copying the list of updates, and
make the HostedObject portion of the implementation simpler.