Refactor ProxyObject to make basic properties reusable without full overhead.
Refactor ProxyObject so the basic state tracking -- basic system
location properties like position, orientation, and mesh -- are stored
in a base class, PresenceProperties, and the version with sequence
numbers in SequencedPresenceProperties. These classes can be used like
lightweight versions of ProxyObjects for parts of the system that only
want that data (e.g. ObjectQueryProcessors, which are only going to
test against that same data again and don't need the functionality of
ProxyObjects).

Also make OrphanLocUpdateManager generic to the type of querier and
use the new SequencedPresenceProperties class instead of it's own
version to store backup data for recent presences.