Use a single stream for manual object host prox queries.
Instead of allocating a substream for each message to the manual
proximity service on the server, allocate a single prox substream and
use it for all messages to the server. This requires some additional
queueing while allocating the stream. This has less overhead and is
simpler since the server doesn't have to worry about unordered
messages. This should result in this implementation having a fixed
stream construction procedure (client's initial request comes in on a
new stream, the space server wouldn't have had to message it before
that), only uses a single stream, and makes messages in each direction
completely ordered.