Allow ObjectQueryProcessor to intercept queries on connection.
When the ObjectHost requests a connection, have it pass the query to
the ObjectQueryProcessor. This makes the ObjectQueryProcessor aware of
the value when the connection is requested and also allows it to
change the value, e.g. if the individual object query should not be
registered, using an aggregate object host query instead. This
approach is implemented in ManualObjectQueryProcessor, whereas the
simple one simply lets the value pass through.