Add enough pass-through of queries to get ManualObjectQueryProcessor to display objects.
Add pass-through (with some minor translation) of query results in
ManualObjectQueryProcessor to display objects. This isn't even close
to how we should be handling those queries, but sufficient to make the
system end-to-end testable.

This required improving tracking of HostedObjectPtrs to perform
callbacks and translation/construction of proximity results from
SequencePresenceProperties objecs, both of which will be useful for
the real querying implementation in the future.