Add FFmpeg build support.
Add FFmpeg library search and make the SDLAudio library depend on it
since we're not going to be able to do much without a way to decode
audio. Test it by performing basic initialization in the
SDL::AudioSimulation class. Update Linux dependencies to get the basic
required dependency. Linux is the only platform currently supported.