Fix audio when format matches target format.
Add some additional insane level debugging for audio. Fix handling of
audio data when it already matches the target format, i.e. it just
needs to be copied over. The size and offset were not being set
properly and we needed to also check the sample rate in addition to
the number of channels and sample format.