Fix dependencies in AddPBJTarget.
Fix the AddPBJTarget to specify the correct command as generating the
pbj.cpp file. This makes the dependencies work properly so that the
correct commands are run for pbj files. This fixes the longstanding
problem where pbj files appeared to sometimes make to need
reprocessing if you re-ran make without changing anything. Also
removes the --delay command from the pbj calls as they are no longer
needed.