Add "locking" support and safe cleanup to Liveness for thread safety.
Add Liveness::Lock to Liveness, which allows you to ensure a Liveness
object remains alive while you hold the lock. This isn't a lock in the
normal sense -- it only blocks the object itself from proceeding with
destruction. This requires that all Liveness objects call letDie()
when they want to invalidate access to themselves, and they may block
waiting for an operation on another thread to finish.