Clean up ProxyEntities properly.
Make ProxyEntities track both when they are active in the renderer (in
the download planner) and when they are active in the result set
(ProxyManager), allowing them to determine when it is safe to destroy
them. Tracking activity in the renderer includes the timeout for
masking quick deletion/addition sequences. Move control of their
deletion to the ProxyEntity itself and expose a method to notify the
OgreSystem when the ProxyEntity has been destroyed.