Fix recursive deadlock issue in AggregateManager.
When a particular path in generateAggregateMeshAsync resulted in
calling generateAggregateMesh while mAggregateObjectsMutex was still
locked, we hit a deadlock because the lock was reacquired. Instead,
this splits generateAggregateMesh into two forms -- one has the
AggregateObjectPtr already and the other doesn't. That path now calls
the one that already has the pointer so it doesn't need to re-lock to
find it.