Fix cleanup of leaf objects/parent pointers in AggregateManager.
AggregateManager wasn't cleaning up objects properly when they were
removed or a child was removed from an aggregate. Partly this is due
to not knowing whether something was a true aggregate or leaf object,
and because the AggregateManager isn't notified of the leaf objects'
removal. We can, however, figure out implicitly whether an object is a
leaf and track it, allowing us to clean things up properly either when
an aggregate is removed or a single child of an aggregate is
removed. This fixes crashes that would occur when objects disconnected
and then the AggregateManager tried to look them up in the
LocationService because it had stale data about them.