Make tracking of visibles in Emerson use data directly where possible.
Change tracking of visibles to use a JSVisibleData interface to track
data. The JSVisibleManager uses JSAggregateVisibleDatas, which combine
information tracked from multiple sources (e.g. visibles from multiple
presences) to present a single source of information for each object
in a space. That data can come from either a ProxyObject or just raw
data, e.g. produced from storage, or default values, e.g. when the
visible is generated by message reception where only the ID is
available.  The JSVisibleManager still listens for updates in order to
determine which source of information is currently the best.

This change shouldn't cause any crashes due to thread
safety, but it isn't properly secured yet, so it is possible in theory
to get reads/writes happening within a ProxyObject at the same time.