Fix getAllObjectReferences to only return active ProxyObjects.
Fix the implementation of getAllObjectReferences to only return active
ProxyObjects rather than any that still have an entry. Without this,
it was possible to (reasonably) expect to get valid shared_ptrs back
from getProxyObject if the SpaceObjectReference was returned, but you
might get NULL if only a weak reference was still available.

This also moves the logic for resetting all proxies into ProxyManager
and just invokes it from HostedObject in order to ensure reset is
called on all objects that are still alive (even those that are not
still active). That way, if the object gets reused, we make sure its
state is properly reset and updates from a new server get applied
after a migration.