Track and report request epoch numbers in HostedObject.
Track requested location update properties in PerPresenceData as a
SequencedPresenceProperties where the sequence numbers stored there
record the epoch number the field was last requested in. This tracks
enough information that we should be able to choose either the field
from the ProxyObject or from the PerPresenceData as the most
up-to-date information. This does not, however, make use of this
information yet.