Handle location update request epochs in space server.
Make LocationService interface expose epochs, add tracking of epochs
to LocationService implementations, and add reporting of epochs in
AlwaysLocationUpdatePolicy. In reporting, we need to decide whether or
not to add the epoch. Currently we do so only when reporting back to
the same object (i.e. to the one requesting changes) or for object
host queries. We would like to filter more for OH queries, but
currently the information about which OH the object is on (and
therefore the information is relevant to it) is not readily accessible
to the update policy.