Make bullet implementation always send updates to ensure objects know about failures.
Make the bullet LocationService always trigger updates for requests,
even if the update isn't applied. We want this so that the object
itself will get notified that the request epoch value has been updated
but the values remain the same, indicating that it needs to revert any
changes related to that change. Theoretically we could do this more
cheaply by adding a way to notify only about epoch changes, which
would only go back to the object itself, but that requires substantial
changes to LocationService, LocationServiceListener, and
LocationUpdatePolicy. This simple approach works fine for now. Fixes