Track per-field request epoch values and ignore old requests.
In the LocationService implementations, use
SequencedPresenceProperties to track per-field request epoch
values. On each request, update using the version that compares
sequence numbers so we only apply newer updates. This handles two
cases properly that a single epoch value can't. First, if an older
update is handled after a newer one but they have distinct fields,
the old update will still be applied and an update sent. Second, when
there aren't distinct fields, older updates will not be applied for
that field. Fixes #476.