Remove old query interface from HostedObject.
This gets rid of some deprecated code that used the old query angle +
max query result interface. Now the only path available is the opaque
query string.

A few changes were required to CSVObjectFactory and
sample scene files, switching from a query angle field to a single,
string-valued query field.

Most changes took place in the JS/Emerson plugin. The C++ side of the
code deals only with the string form of the query, making it very
flexible. A small amount of parsing/encoding code is added to
presences, assuming for now that the only format is a JSON-encoded
angle + max result format. However, the raw string is also made
available and mutable so we can easily support other query
formats. The rest of the changes in the plugin handle the new format
for storage, including compatibility with the previous format so we
can still restore old objects.