Use synchronous onCreateProxy when adding proxy for connected presence.
We explicitly call onCreateProxy when a presence is connected in the
Emerson plugin in order to seed the objects information because the
proxy is created before there is a chance to subscribe for
notifications. When we switched to a multithreaded model this call
remained the same, but internally the callback posts a handler to the
script's strand. However, we want the effects of the handler to occur
immediately and we were already in the correct strand in this case, so
convert it to a direct call to the internal handler. Fixes the
createPresence Emerson unit test.