Add stream-based ModelsSystem::convertVisual interface.
Add a version of convertVisual which works with streams instead of
using filenames. This allows either in-memory or to-disk
conversions. Deprecate the old version as it is trivially accomplished
using an ofstream.

Because OpenCOLLADA currently only supports saving to a file (despite
using a generic stream interface to write to, it only has a
constructor taking a filename), add a utility for generating temp file
names and use it to implement the in-memory conversion using the old
file-based version.

Convert the two uses of convertVisual to use this new interface:
SaveFilter just uses an ofstream since it always targets a file and
AggregateManager also uses an ofstream but will soon be converted to
use a stringstream so it can then be passed into the transfer system
for upload.