Cache compiled Emerson.
When doing an import in the Emerson plugin, try to read from a cache,
stored as a replicated tree under a temp directory (using absolute
paths to ensure no conflicts). The regular path takes the path to the
cache as well and saves the compiled JS there. As long as the cached
JS is newer than the source file the cached JS can be used instead.

Especially when starting many objects, this significantly improves
startup performance because a lot of files are otherwise unnecessarily
recompiled for each object. A simple test with the default scene shows
about 20x speedup. Because timestamps are checked, this does not
change the behavior of repeated imports of the same file.