Remove scan through system.presences for each connected presence.
Previously, to decide whether to add a presence to the
system.presences array, we'd scan through the entire array, convert
the presences to strings, and compare them. This became a bottleneck
if you tried to connect a lot of presences from a single entity
because it resulted in a quadratic operation (N checks through an O(N)
array) and dropping to C++ and doing the conversion to a string wasn't
particularly fast either.

This patch just changes the way we decide to add presences to the
array, using the system object's _selfMap to filter presences that are
already there. Loading about 2000 objects used to take about 5
minutes and now it takes about 8 seconds.