Add query processing for ManualQueryProcessor.
Add ObjectQueryHandler for doing query processing locally based
on the output OHLocationServiceCache from server queries. Instantiate
one per server when queries are registered and have it deliver the
results. This gets enough pieces connected to get results back to the
querier, allowing a client to display the world.

Some issues still exist: location updates are not provided, updates to
queries are not yet handled, objects aren't loaded into the query
processor in the same structure they are received from the server (we
need parent information and a new way to add objects for that), and in
order to get these results we always fully refine the results provided
by the server.