Check connection validity before retrying session messages.
Previously we would always retry sending session messages in
sendSessionMessageWithRetry because we only got a boolean
back. However, the send could fail either because of a temporary full
queue or because the object host connection was closed. In the latter
case, we could get stuck in an infinite loop.

This adds methods to check the state of the connection
(ObjectHostConnectionManager::validConnection), allowing that code to
only retry if the connection is still valid.

Fixes #483.