A few fixes to give accurate profiling data.
Fix OgreRenderer to not perform an extra post, instead handling poll()
calls immediately since its parent, TimeSteppedSimulation, already
guarantees it runs in the correct strand. This fixes the accounting of
time for OgreRenderer.

Fix the name passed to PollingService by SessionManager so it isn't
reported as 'Space Server', which doesn't make any sense on the OH.