Tag and track post()s to IOServices and IOStrands.
Add a tag parameter to IOService and IOStrand post() calls to track
the type of callbacks on their queues. Add callbacks for all calls
currently made, but allow a default NULL tag so it's not required to
add tags.

Tags are strings, represented as const char*'s. This makes it
efficient to hash and count them. When a report is requested, we make
sure we aggregate all equivalent strings, then sort them and add them
to the report.