Fix StorageTestBase::testAtomicWriteErase to not depend on previous state of the database.
Previously testAtomicWriteErase depended on the key 'k' already being
in the database, but tests executed before it didn't ensure this. This
could cause the test to succeed or fail depending on the previous
state of the database, and, importantly, depending on whether the
tests had previously been run. Instead, it now just writes, then
erases, that key, ensuring it is gone.

While fixing this, fix the typo in the method names and add these
tests to the SQLiteStorageTest.