Ensure swapping objects generates addition/removal events in the correct order.
Previously objects were added/removed from query handlers immediately
when moving between static and dynamic query handlers. This could
cause an object to get lost for a querier because the query handler it
removed from was ticked *after* the one it was added to, causing what
appeared to be a redundant addition followed by a removal. This
changes the way we handle the swap by processing all removals before a
tick and additions afterward so that the addition events are
guaranteed to be generated after the removal events.