Make LocUpdates transparently convert to local time.
Remove the *WithLocalTime methods from LocUpdate, instead adding
comments to note that LocUpdate methods should always return values
with local time. This is simpler for consumers of LocUpdates since
they don't have to do the conversion and isolates all time conversion
code to be as close to the reception of the messages as possible.

This change also fixes incorrect behavior with manual queries where
time conversion was being performed twice, resulting in incorrect
locations for moving objects.

Finally, since this required that OrphanLocUpdateManager use the
ObjectHost to do the conversion, move that and
PresencePropertiesLocUpdate into liboh. This fixes some liboh includes
that were in libproxyobject.