Add refreshing of TTLs for uploaded aggregate meshes.
Track the expiration time for each uploaded aggregate and use a Poller
to check for aggregates which are still alive and nearing
expiration. Send keep alive requests for those aggregates with new
TTLs. This is currently a simple brute force search, but it only
happens every 30 seconds. A more efficient event-based approach might
be necessary with a large number of aggregates.