Use reported canonical subfile paths when parsing Collada and when uploading aggregates.
Extract canonical subfile paths from headers (e.g. Subfile-0-Name and
Subfile-0-Path) and replace texture references in the parsed Collada
with full, canonical URLs.

In aggregate manager, rework the generation of subfiles to use these
URLs, or work for non-URL textures. The keys for subfiles are always
the path basename (i.e. just the filename, no parent directories) and
the value is the CDN-style resource path (the name of the asset on the
CDN rather than the exact URL used for download).

Similar to generating the URL for aggregate meshes from the path
reported by the upload process, this requires switching around version
and filenames in URLs -- first when placing URLs in the Meshdata so
that display (i.e. download) will work properly then back to the
original when using the URLs for specifying existing files for
aggregate textures.