Cananicalize FILE_EXE path on Mac.
The function used to get the executable's path on Mac can return
paths with relative path components (. and ..). These can interfere
with assumed path offsets.

Add a method for removing these elements (.. only when specifically
requested since doing so is not always safe. Apply this to the
FILE_EXE path generated on Mac.

This addresses an issue where cppoh would just hang on start if
started on Mac with ./ as the first part of the executable path.