Add a simple HTTP implementation of Commander.
Add a new plugin that listens for commands via HTTP requests and
dispatches them. Set this as the default for space and implement
single command in Server that returns the number of objects as a
simple test.

The HttpCommander class is relatively simple, but relies on a new
HttpServer class that implements a simple HTTP server. It's built on
the http_parser library and essentially provides the half of HTTP
compared to HttpManager.