Change OH::Storage interface to return success or cause of error.
Replace the bool success parameter to OH::Storage callbacks with an
enum that indicates success or the reason for an error. This should
only contain a small, fixed, implementation-independent set of
errors. The two currently included are LOCK_ERROR, indicating an error
with access to the underlying database, and TRANSACTION_ERROR
indicating the transaction failed to apply cleanly (e.g. missing key,
failed comparison). This allows us to expose some aspects of the
underlying database connection abstractly.