Fix tests in build system.
CMake doesn't like it if you use the same build directory for two
subdirectories, which we were doing for json-spirit. There's a policy
to fix this, CMP0013, but apparently there was still some sort of bad
interation that caused the tests and test targets to not work, despite
CMake definitely reaching those commands.

The only reason for using the same directory was that on Windows the
dynamic library couldn't be found when running in-tree if a separate
directory was used. This switches back to using a separate directory,
but uses a static json-spirit library on Windows and a shared library
on other platforms.