Add options support for test binary.
Support options in the test binary. Pass the options through and
initialize and parse for each test suite. Since some suites muck with
options (e.g. the options test), we need to make sure we're in a clean
state for each test.

This change is especially helpful in debugging test failures since it
allows the log levels to be adjusted.