Add lease support to CassandraStorage implementation.
Add lease support to CassandraStorage implementation. This looks
structurally similar to the SQLite implementation where we acquire
leases on demand when a transaction is processed and then maintain a
queue of renewal requests.

The implementation of actually acquiring and maintaining leases is
based on a wait-queue algorithm, although we don't actually wait in
the queue: we just determine if we win if there is contention and then
succeed or fail, which greaty simplifies the implementation. This
requires a new column family for leases (which are kept separate from
the actual persistence data, and which uses a 'Standard' rather 'Super'
column family).