Add HTTP command helper classes and most basic test.
Add helper classes for tests using the HTTP command
interface. HttpCommandTest provides helpers for setting up nodes that
need the http command interface turned on and a simple interface for
executing commands, possibly retrying if necessary.

SpaceHttpCommandTest subclasses this for tests that only involve the
space server. It's similar to CSVTest in strcture, but only starts up
the space.

The one test that's added executes the 'meta.commands' command, which
is always available and makes sure it got a valid response.