Use index of HostedObjects by internal ID when stopping ObjectHost.
The old method of iterating over all presences and
generating a unique list of HostedObjects is both inefficient and
misses HostedObjects which have closed all their connections but not
shutdown. Now that we have the map by internal IDs, use that instead,
which should catch all HostedObjects and is simpler.