Add prox commands to OH manual query processor.
Add prox commands that have been added to the space servers to the OH
manual query processor. The setup is a bit trickier because we can
have lots of query processors. The naming has the node idea, the type
of queriers, and then the type of objects queried (static or dynamic).