Service SST streams after acks so waiting packets are pushed out.
Previously we were not scheduling any servicing of SST stream. This
probably worked well back when we were polling SST streams, but now
you can end up in a state where your next servicing is many seconds in
the future (e.g. because you're waiting 2x RTT before triggering a
servicing to resend data, but your RTT can start out quite high).

This just schedules servicing if any packets are outstanding and we
actually acked some packets, freeing up room in the transmit
window. This is probably a bit of overkill since it schedules
servicing for every ack, but that's a larger problem with how
servicing is scheduled in the SST code.

This fixes slow throughput issues with the proximity code, where a lot
of data is dumped on the stream at once right when the stream is
connected and it would trigger this issue because the initial RTT was
hight (and stream servicing schedule therefore slow).