Coalesce all events from first iteration of object queries.
In LibproxProximity, i.e. for standard solid angle object queries,
coalesce all events up until the end of the first tick event into a
single event (or sequence of events). This skips all the intermediate
refinement, just reporting an initial set of objects. This saves a
bunch of effort and isn't so misleading to clients, which sometimes
have to wait for refinements to stream in. Instead, this just reports
objects that, at the time of the query registration, really are in the
result set.

This shouldn't have any effect on queries that don't use
aggregates. The coalescing code is a generic method in
LibproxProximityBase since it might be useful in other places.