Fix use of invalidated iterator in OgreRenderer.
We use a std::list of moving entities so it can be modified while
iterating over it, but we were incrementing the iterator after the
Entity was ticked. Since the entity can be removed from the list
during that call, we need to make sure we've advanced the iterator
before invoking tick().

Fixes #500.