Make speed changes apply immediately and handle interleaved speed changes and start/stop movement.
Track active movement requests and undo and then reapply them when
a speed change is requested. This also requires changing moveAndRotate
to allow for intermediate updates that aren't actually applied so it
doesn't interfere with animations (by stopping them because an
intermediate step of this process is to put the object back into a
static state).

Fixes #503.