Fix error in ThreadSafeQueueWithNotification that could make it miss a notification.
The code was checking before pushing if the queue was empty, but it
could go empty between the check and actually pushing since this check
wasn't performed when the queue was locked. This change makes
ThreadSafeQueue's push methods return the size of the queue after
pushing the items on, but gets this value while the queue is
locked. Using this value we can tell if the queue was empty when we
pushed items onto it, thus getting the notifications correct.

Also fixes #499, which was caused by losing a notification in the
SQLite storage implementation, causing processing to stop and the
thread waiting for the result to hang.