Add checks for expected functionality in PresenceEventsTest.
The previous fix was necessary because we didn't properly setup
ProxManager to ensure callbacks will be able to use the functionality
we expect in them. This adds a couple of simple tests to the
Emerson PresenceEventsTests to check that this functionality is
available. Currently the tests are whether system.self is set and
whether we can call getProxSet.